requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
Django==3.0.3
django-extensions==2.2.8
django-rest-swagger==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.1
pluggy==0.13.1
psutil==5.7.0
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wincertstore==0.2
zipp==2.2.0
